+++ CHANGELOG.md
@@ -6,3 +6,5 @@
     [JENKINS-19942](https://issues.jenkins-ci.org/browse/JENKINS-19942)
   * register input files for the localizer task to avoid that the task is skipped when the inputs change
     [JENKINS-24298](https://issues.jenkins-ci.org/browse/JENKINS-24298)
+  * added pluginFirstClassLoader attribute
+    [JENKINS-24808](https://issues.jenkins-ci.org/browse/JENKINS-24808)